Extract `EditorTheme` to its own file
