+++ src/vs/editor/browser/view/viewPart.ts
@@ -5,7 +5,7 @@
 
 import { FastDomNode } from 'vs/base/browser/fastDomNode';
 import { RenderingContext, RestrictedRenderingContext } from 'vs/editor/browser/view/renderingContext';
-import { ViewContext } from 'vs/editor/common/viewContext';
+import { ViewContext } from 'vs/editor/common/viewModel/viewContext';
 import { ViewEventHandler } from 'vs/editor/common/viewEventHandler';
 
 export abstract class ViewPart extends ViewEventHandler {
