+++ src/vs/editor/browser/viewParts/currentLineHighlight/currentLineHighlight.ts
@@ -7,7 +7,7 @@ import 'vs/css!./currentLineHighlight';
 import { DynamicViewOverlay } from 'vs/editor/browser/view/dynamicViewOverlay';
 import { editorLineHighlight, editorLineHighlightBorder } from 'vs/editor/common/core/editorColorRegistry';
 import { RenderingContext } from 'vs/editor/browser/view/renderingContext';
-import { ViewContext } from 'vs/editor/common/viewContext';
+import { ViewContext } from 'vs/editor/common/viewModel/viewContext';
 import * as viewEvents from 'vs/editor/common/viewEvents';
 import * as arrays from 'vs/base/common/arrays';
 import { registerThemingParticipant } from 'vs/platform/theme/common/themeService';
