+++ src/vs/editor/browser/viewParts/viewZones/viewZones.ts
@@ -9,7 +9,7 @@ import { IViewZone, IViewZoneChangeAccessor } from 'vs/editor/browser/editorBrow
 import { ViewPart } from 'vs/editor/browser/view/viewPart';
 import { Position } from 'vs/editor/common/core/position';
 import { RenderingContext, RestrictedRenderingContext } from 'vs/editor/browser/view/renderingContext';
-import { ViewContext } from 'vs/editor/common/viewContext';
+import { ViewContext } from 'vs/editor/common/viewModel/viewContext';
 import * as viewEvents from 'vs/editor/common/viewEvents';
 import { IViewWhitespaceViewportData } from 'vs/editor/common/viewModel';
 import { EditorOption } from 'vs/editor/common/config/editorOptions';
