+++ src/vs/editor/common/viewModel/viewContext.ts
@@ -7,34 +7,7 @@ import { IEditorConfiguration } from 'vs/editor/common/config/editorConfiguratio
 import { ViewEventHandler } from 'vs/editor/common/viewEventHandler';
 import { IViewLayout, IViewModel } from 'vs/editor/common/viewModel';
 import { IColorTheme } from 'vs/platform/theme/common/themeService';
-import { ColorIdentifier } from 'vs/platform/theme/common/colorRegistry';
-import { Color } from 'vs/base/common/color';
-import { ColorScheme } from 'vs/platform/theme/common/theme';
-
-export class EditorTheme {
-
-	private _theme: IColorTheme;
-
-	public get type(): ColorScheme {
-		return this._theme.type;
-	}
-
-	public get value(): IColorTheme {
-		return this._theme;
-	}
-
-	constructor(theme: IColorTheme) {
-		this._theme = theme;
-	}
-
-	public update(theme: IColorTheme): void {
-		this._theme = theme;
-	}
-
-	public getColor(color: ColorIdentifier): Color | undefined {
-		return this._theme.getColor(color);
-	}
-}
+import { EditorTheme } from 'vs/editor/common/editorTheme';
 
 export class ViewContext {
